Add an id field to Person class
Each person should have their own ID, it must be
unique and static (which does not change after
editting). The simplest implementation is to
used a static int field.

Add a static and a private field to Person
class. Update the logic of storing and reading
data from storage. Update the logic of edit
a person.